requirements.txt
python-dotenv
google-auth
google-cloud-core
google-cloud-aiplatform
google-cloud-texttospeech
google-cloud-translate
pydub
networkx
spacy
pandas
numpy
matplotlib
altair
requests
anthropic                      #for querying the Anthropic models hosted on Google Cloud
pytest
pysnooper
librosa                         #for speeding up dialogue
soundfile                       #for speeding up dialogue